backend/app/ai_integration.py
from typing import Dict, Any, Optional
import os
from google.cloud import aiplatform
from google.oauth2 import service_account
import json

class AIIntegration:
    def __init__(self, credentials_path: Optional[str] = None):
        """
        Initialize the AI integration with Google's Gemini API.
        
        Args:
            credentials_path: Path to Google Cloud credentials JSON file
        """
        if credentials_path:
            credentials = service_account.Credentials.from_service_account_file(
                credentials_path
            )
            aiplatform.init(credentials=credentials)
        else:
            # Use default credentials
            aiplatform.init()

        self.model = aiplatform.TextGenerationModel.from_pretrained("gemini-pro")

    def summarize_code(self, code: str, language: str, node_type: str) -> Dict[str, Any]:
        """
        Generate a summary for a code snippet.
        
        Args:
            code: The code to summarize
            language: Programming language of the code
            node_type: Type of node (function, class, etc.)
            
        Returns:
            Dictionary containing the summary and metadata
        """
        prompt = self._create_summary_prompt(code, language, node_type)
        
        try:
            response = self.model.predict(prompt)
            summary = response.text
            
            # Parse the response into structured format
            return self._parse_summary(summary, code, language, node_type)
        except Exception as e:
            return {
                "error": str(e),
                "summary": "Failed to generate summary",
                "metadata": {}
            }

    def _create_summary_prompt(self, code: str, language: str, node_type: str) -> str:
        """Create a prompt for code summarization."""
        return f"""Analyze this {language} {node_type} and provide:
1. A single-sentence summary of its primary purpose
2. A list of its parameters/inputs
3. What it returns/outputs
4. Any important side effects or dependencies

Code:
{code}

Format your response as JSON with the following structure:
{{
    "summary": "single sentence summary",
    "parameters": ["param1", "param2", ...],
    "returns": "description of return value",
    "side_effects": ["effect1", "effect2", ...],
    "dependencies": ["dep1", "dep2", ...]
}}"""

    def _parse_summary(self, summary: str, code: str, language: str, node_type: str) -> Dict[str, Any]:
        """Parse the AI response into a structured format."""
        try:
            # Try to parse as JSON
            data = json.loads(summary)
            return {
                "summary": data.get("summary", ""),
                "parameters": data.get("parameters", []),
                "returns": data.get("returns", ""),
                "side_effects": data.get("side_effects", []),
                "dependencies": data.get("dependencies", []),
                "metadata": {
                    "language": language,
                    "node_type": node_type,
                    "code_length": len(code)
                }
            }
        except json.JSONDecodeError:
            # If parsing fails, return the raw summary
            return {
                "summary": summary,
                "parameters": [],
                "returns": "",
                "side_effects": [],
                "dependencies": [],
                "metadata": {
                    "language": language,
                    "node_type": node_type,
                    "code_length": len(code),
                    "parse_error": True
                }
            }

    def generate_embeddings(self, text: str) -> Dict[str, Any]:
        """
        Generate embeddings for text using Gemini.
        
        Args:
            text: Text to generate embeddings for
            
        Returns:
            Dictionary containing the embeddings
        """
        try:
            # TODO: Implement embedding generation
            # This is a placeholder. In production, we would:
            # 1. Use Gemini's embedding model
            # 2. Cache results
            # 3. Handle rate limiting
            return {
                "embeddings": [],
                "error": "Embedding generation not implemented"
            }
        except Exception as e:
            return {
                "embeddings": [],
                "error": str(e)
            } 