backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
gitpython==3.1.40
tree-sitter==0.20.4
networkx==3.2.1
redis==5.0.1
google-cloud-aiplatform==1.36.4
python-dotenv==1.0.0
pydantic==2.4.2
pytest==7.4.3
httpx==0.25.1 